stock-trading-capstone/requirements.txt
# Основні бібліотеки
pandas>=1.5.0
numpy>=1.24.0,<2.0  # Сумісність з PyTorch та іншими пакетами
matplotlib>=3.6.0
seaborn>=0.12.0

# Веб-додаток та візуалізація
streamlit>=1.28.0
plotly>=5.17.0

# Дані
yfinance>=0.2.0

# Машинне навчання
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Прогресс-бари та утиліти
tqdm>=4.65.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.25.0

# Додаткові пакети для статистики
scipy>=1.11.0
statsmodels>=0.14.0

# Для Prophet (опціонально)
prophet>=1.1.4

# AutoML (опціонально)
auto-sklearn>=0.15.0

# 🌅 Sundial Foundation Model Dependencies
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.16.0

# Підтримка Apple Silicon
# (встановлюйте через install_apple_silicon.py)
